add --load-dtd option
This option tells itstool to load external DTDs when
parsing the document to be translated.  This prevents
errors when the document includes entity references defined
in those DTDs.

Note that externally-defined entity refs still cannot be used
in translated strings in the PO files.

Also note that this adds test cases that require either
network access or updating the local XML catalog to
including the DocBook DTDs.

Signed-off-by: Galen Charlton <gmc@esilibrary.com>