Don't re-lookup ITS params for every single rule
We have to create a new XPath context for every rule, and we have
to freshly apply any ITS params to that context. But what we were
also doing is looking thru the children of the parent ITS rules
element for param element each time. This was inefficient. Looking
for param elements only once per ruleset gives a small but measurable
speed boost.