Cache computed values for translate, localefilter, and locnote
If these aren't explicitly set for a node with a selector or local markup,
then we have to chain up, often all the way to the root. But we were doing
this over and over again, chaining up across the same nodes. This change
stores the computed value, so chaining up usually hits that cache. I was
careful to clear the cache if new rules are applied.