tests: use same Python executable
Instead of hardcoding to 'python3' or 'python2' the executable to use to
run itstool_test, use the same Python executable that was used to run
the test suite. Fallback to 'python3' in the rare cases when
sys.executable is not set.

This makes it possible to execute the test suite by running run_tests.py
using the desired Python executable (e.g. of a specific installation or
version).