Fix regex processing with python 3.12
I see an error spam when building Epiphany with itstool 2.0.7-5.fc39:

```
[42/297] Generating help/gl/help-epiphany-gl with a custom command
/usr/bin/itstool:236: SyntaxWarning: invalid escape sequence '\s'
  if re.sub('\s+', ' ', text).strip() != '':
/usr/bin/itstool:334: SyntaxWarning: invalid escape sequence '\s'
  message = re.sub('\s+', ' ', message).strip()
/usr/bin/itstool:472: SyntaxWarning: invalid escape sequence '\s'
  return re.sub('\s+', ' ', self.locnote).strip()
/usr/bin/itstool:474: SyntaxWarning: invalid escape sequence '\s'
  return '(itstool) link: ' + re.sub('\s+', ' ', self.locnoteref).strip()
/usr/bin/itstool:905: SyntaxWarning: invalid escape sequence '\<'
  regex = re.compile('(.*) \<(.*)\>, (.*)')
/usr/bin/itstool:940: SyntaxWarning: invalid escape sequence '\s'
  if re.sub('\s+', '', prevtext) == '':
/usr/bin/itstool:1484: SyntaxWarning: invalid escape sequence '\.'
  _locale_pattern = re.compile('([a-zA-Z0-9-]+)(_[A-Za-z0-9]+)?(@[A-Za-z0-9]+)?(\.[A-Za-z0-9]+)?')
```

Problem is what were intended to be regex escapes are instead being
treated as invalid python string escapes. To get a regex escape, we need
to either escape the escape, i.e. use `'\\s+'` instead of `\s+'`, or --
better -- use raw string notation, `r'\s+'`.

There's a warning about this in the python 3.12 release notes,
https://docs.python.org/3.12/whatsnew/3.12.html ("A backslash-character
pair that is not a valid escape sequence now generates a SyntaxWarning,
instead of DeprecationWarning.")